Clean up TTL/class order, other stuff
Also puts variables handling ($ORIGIN, $TTL) into the parser
instead of the grammar. Doing syntax tree navigation inside
what's basically a fancy lexer is not very clean. This means
we lose a bunch of tests in the spec, but we implicitly check
that through RR checks.

A bunch of refactoring in the parser itself, makes RR-specific
implementations shorter, cares for variable states, etc.